Created the Graphics Service and linked it to the Service Locator
Created the Graphics Service for the game which is responsible for displaying the game window and rendering the different graphics entities in that window.

Created the Graphics service by creating the GraphicService header file and source file and declaring and defining the necessary pointers and methods.

Linked the GraphicService header file to the ServiceLocator header file and updated the ServiceLocator header and source files accordingly in order to call GraphicService functions.

Made necessary changes to GameService header and source files in order to call the GraphicService functions through the ServiceLocator.