Created the Event Service and linked it to the Service Locator
Event Service is responsible for checking user input interactions/events like keyboard button presses and mouse clicks and then performing certain actions based on the input interactions.

The Event Service created in this code base checks whether the user has pressed the "Esc" key and then will close the game window and terminate the program if the key has been pressed.

Created the EventService header and source file and declared and defined the necessary methods, an object of the Event class provided by SFML and a pointer to object of the RenderWindow class to store instance of game window.

The methods defined included a check for whether a key has been pressed,  whether the key pressed is the "Esc" key and a method for closing the game window if the above methods returned true.

The Event Service was linked to the ServiceLocator by including the EventService header file in the ServiceLocator header file.

The ServiceLoacator header and source files were updated accordingly by  creating a pointer to an instance of the EventService class and creating a getter method to get pointer to instance of EventService in the ServiceLocator header file and by initializing the EventService pointer and calling the EventService methods in the ServiceLocator source file.

The update method in the GameService source file was then updated to call the EventService method for processing the events as mentioned earlier.