Created the Player Service and linked it to Service Locator
The Player Service is responsible for rendering the player ship and allowing left and right movement of the ship in the game window.

Created the PlayerService header and source files.

In the PlayerService header file the required variables were initialized and appropriate methods were declared to implement rendering of the ship in the game window and the movement of the ship.

Created functions in the EventService for checking whether right key or left key is pressed for player movement.

Created a pointer to instance of EventService to implement player movement by calling the methods created in the EventService for checking whether left key or right key is pressed and implemented the methods in the PlayerService source file.

The PlayerService was linked to the ServiceLocator by including the PlayerService header file in the ServiceLocator header file.