Created the Time Service and linked it to the Service Locator
The Time Service is responsible for calculating the amount of time elapsed between consecutive frames. This calculated time called delta time can then be multiplied by the movement speed in the Player Service to make the player ship movement time dependent instead of frame dependent hence ensuring consistent player movement and thereby gameplay experience across all hardware.

Created the TimeService header file and source file. Declared the necessary variables and methods to calculate and store delta time in the header file. The methods were defined in the source file.

Linked the Time Service to the Service Locator by including the TimeService header file in the ServiceLocator header file and making necessary changes in the ServiceLocator header and source files.

The PlayerService was updated accordingly to make the player movement time dependent. The change was multiplying the calculated delta time with the player movement speed to implement player movement.

In the PlayerService the the method for player movement was replaced with two methods for left player movement and right player movement.

In the GraphicService a frame rate limit of 60 was set.