Created the Player Model class
The Player Model class is responsible for managing i.e. updating and returning all the data related to the player such as the player score, player state(alive, dead), player position.

Created the PlayerModel header file and source file. Created the Player Model class and declared and defined all the required variables and methods to store and update the player data in the header and source file.

Replaced the bool variable to check whether player is alive or dead with an enum class which stores player state (alive, dead) in the PlayerModel.