Created the Player View and Player Controller
The Player View is responsible for rendering the player ship onto the game window and showing the changing positions of the ship in the game window.

Created the PlayerView header and source files and declared and defined the required variables and functions to display the player ship in the game window and set its updated positions.

The Player Controller is responsible for processing player inputs/interactions and communicating with/calling the necessary functions of the Player View and the Player Model.

Created the PlayerController header and source files.

Linked the Player Model and Player View to the Player Controller by including their header files in the PlayerController header file.

Declared and defined the necessary methods and pointers to instance of the Player Model and Player View classes in the Player Controller header and source files.

The Player Controller will check whether left or right key has been pressed and will update the position of the player by calling the appropriate Player Model function. The appropriate Player View function will be called to render the player ship in the game window and set its updated positions.

Linked the Player Controller to the Player Service by including the PlayerController header file in the PlayerService header file.

The PlayerService header and source files were updated accordingly by removing all the previously defined methods in Player Service which are now being handled by Player Model, Player View and Player Controller and created a pointer to instance of Player Controller class to call Player Controller methods.