Resolved Circular dependency problem by replacing header file inclusions with forward declarations of required classes.
Replaced the inclusion of PlayerView and PlayerModel header files in PlayerController header file with declarations of PlayerView class, PlayerModel class and PlayerState enum class and moved the header file inclusions to the PlayerController source file.

Replaced the inclusion of PlayerController header file in the PlayerView header file with PlayerController class declaration and moved the header file inclusion to the PlayerView source file.

Replaced the inclusion of PlayerController header file in PlayerService header file with declaration of PlayerController class and moved the header file inclusion to the PlayerService source file.

Reorganized the folder and file structure by making seperate folders for each service and updated the header file inclusions accordingly.