Created namespaces for different services
Created individual namespaces for each service and the Service Locator.

Grouped all the declarations in each header file of each service in the respective namespaces.

Grouped all definitions in all the source files of the services and the Service Locator in the respective namespaces.

For Player all the declarations and definitions in PlayerService, PlayerController, PlayerModel and PlayerView header and source files were grouped under a single namespace.

Used the namespaces for services in the namespace for Service Locator in ServiceLocator header and source file.

Used the namespace for Service Locator wherever Service Locator instance is being used in the service source files.

Used the namespace for Game Service in main function of the code base.