Created Enum class for different game states
Created an Enum class for Boot, Main menu and Gameplay states of the game in the GameService header file.

Declared and defined a variable to store current state of the game and functions to set a new game state and get current game state in the GameService header and source files.