Created the UI Service and Main Menu Controller
The UI Service is responsible for managing and connecting to all the different game state controllers that manage and display different elements of the UI in the game window.

Main Menu Controller is responsible for displaying the elements of the main menu in the game window and managing user interactions with the elements of the main menu.

Created the UIService header and source files. Created UIService class in the UI namespace in the header file.

Created the MainMenu Controller header and source files. Created the MainMenuController class in the MainMenu namespace nested in the UI namespace in the header file. Declared and defined pointer to instance of RenderWindow class and the neceassary methods to display main menu elements and manage user interaction with main menu elements. Included the header files and used namespaces of Graphic Service, Event Service, Service Locator and Game Service in the MainMenuController source file.

Linked the Main Menu Controller to the UI Service by including the MainMenuController header file in the UIService header file.

Declared and defined the pointer to instance of MainMenuController class and the necessary methods to manage and connect to game state controllers in the UIService header and source files.

Linked the UIService to the ServiceLocator by including the UIService header file in the ServiceLocator header file. Declared and defined a pointer to an instance of UIService class and a function to get UIService instance in the ServiceLocator header and source files and called the required functions of the UIService class in the ServiceLocator source file.