Created the buttons for the Main Menu and used image as background for the game window
Created texture and sprite objects for the background and all the buttons of the main menu in the MainMenuController header file.

Initialized the variables to store the texture paths and declared and defined necessary functions to load textures, set the textures to the sprites, set the scale/size of the sprites and positions of the button sprites in the MainMenuController header and source files.

Increased the game window size to fit the entire screen by making the necessary changes in the GraphicService header and source files.

Made changes to the initial player position, the player movement speed and the left most boundary and right most boundary for player movement in the PlayerModel header file.

Called the MainMenuController render function from the UIService source file to display the main menu buttons and background image in the game window.