Implemented button interactions for the main menu
Set the initial Game State to Main Menu using the GameState enum class and setGameState function created previously in the GameService source file.

Made changes to the UI Service source file by adding a switch case in the UIService render and update functions to call MainMenuController update and render functions only if the Game State is Main Menu.

Declared and defined functions to check whether left mouse button or right mouse button is being clicked in the EventService header and source files.

Declared and defined functions to process button interactions for the main menu in the MainMenuController header and source files.

Processed Main Menu button interactions in the MainMenuController source file by using the mouse position function and a function to get the bounds of each button sprite provided by the sfml library and the button click check functions created in the Event Service to check which button the mouse is hovering over and whether the left mouse button has been clicked and then changing the Game State accordingly.