Created an enum class for button states (PRESSED, HELD, RELEASED)
Fixed an issue where the main menu instructions button click was being registered multiple times just by holding left mouse button click over the main menu  instructions button.

Implemented left and right player movement using the 'A' and 'D' keys.

Fixed the issue by creating an enum class in the EventService header file that stores button states (PRESSED, HELD, RELEASED) and using the enum to accordingly change the state of mouse button click to "HELD" if the mouse button was being held down so it would not be registered as "PRESSED" for the entire time the mouse button is being held and would register the instructions button click only once instead of multiple times.

Declared and defined the functions to change the mouse click button press and key press states and according to enum states changed the existing functions to check the states of the mouse click and key press in the EventService header and source files.

Declared and defined functions to check for 'A' and 'D' keyboard key presses in the EventService header and source files.

Called the necessary functions of EventService to check the button states of left and right arrow key press and the 'A' and 'D' key press in order to process player movement in the PlayerController source file.