Created the Enemy Service and the Enemy Model, View and Controller classes
The Enemy Service is responsible for connecting with the enemy controller, model and view to call the necessary functions to implement the enemy rendering and movement on the game window.

Created the EnemyService header and source files. Created the Enemy namespace and created the EnemyService class and declared and defined necessary functions and pointer to instance of Enemy Controller class in the EnemyService header and source files. to call the functions of the EnemyController in the EnemyService source file.

Created the Enemy Controller header and source files. Forward declared the EnemyModel and EnemyView classes in the EnemyController header file and declared and defined necessary functions and pointers to instances of EnemyModel and EnemyView classes in EnemyController header and source files to call their functions in the EnemyController source file.

Created the EnemyModel header and source files. Declared and defined appropriate fucntions to set and return enemy position.

Created the EnemyView header and source files. Declared and defined texture and sprite variables and necessary functions to display the enemy in the game window in the header and source files.

Linked the Enemy Service to the Service Locator by including the EnemyService header file in the ServiceLocator header file.

Linked the Enemy Model and Enemy View to the Enemy Controller by including the EnemyModel and EnemyView header files in the EnemyController source file.