Implemented the Enemy movement
Implemented enemy movement which allows enemy to move from left most position to right most position then move down from the right most position by a specified amount then from there move to the left most position then move down again and then repeat the same cycle.

Created an enum class for Enemy movement directions (LEFT, RIGHT, DOWN) in the EnemyModel header file, declared and defined the necessary functions to set and return the Enemy movement direction in the header and source files and created variables for left most and right most position of enemy movement, the distance by which the enemy moves down and the enemy movement speed in the header file.

Declared and defined necessary functions in the EnemyController header and source files to implement movement of the Enemy as per the directions specified in the MovementDirection enum.