Created the Gameplay Service and the Gameplay Controller and View classes
The Gameplay Service is responsible for connecting with the GameplayController which in turn connects with the GameplayView for processing and showing changes in the state of the game while it is being played.

Created the Gameplay Service, Controller and View header and source files. Created the Gameplay namespace and the GameplayService, GameplayController and GameplayView classes in their respective header files.

Declared and defined the pointer to GameplayController class and the necessary functions in the GameplayService header and source files to call the GameplayController functions in the source file.

Declared and defined the pointer to instance of GameplayView class and the necessary functions in the GameplayController header and source files to call the GameplayView functions in the source file.

GameplayView is currently responsible for displaying the background image in the game window for when the game is being played.

Declared and defined the texture and sprite variables for the background image and the necessary functions to display the background image in the game window in the GameplayView header and source files.

Linked the GameplayView to the GameplayController by including the GameplayView header file in the GameplayController source file.

Linked the Gameplay Controller to the Gameplay Service by including the GameplayController header file in the GameplayService source file.

Linked the Gameplay Service to the Service Locator by including the GameplayService header file in the ServiceLocator header file and declaring a pointer to instance of GameplayService class in the ServiceLocator header file to define the pointer and use it to call the GameplayService functions in the ServiceLocator source file.