(Personal Code Improvement) Used the background image sprite for Main Menu in the Gameplay View
Declared and defined function to return the background sprite in MainMenuController header and source files.

Declared and defined function to get MainMenuController instance in the UIService header and source files which can be used to call the MainMenuController function to get background sprite.

Retrieved the background sprite defined for Main Menu in the GameplayView source file and displayed it in the game window.

These changes were done for the purpose of reducing code redundancy by removing the variables and functions for loading background image texture into background image sprite and scaling and displaying the background image sprite in the game window from the GameplayView header and source files as it had already been implemented in the MainMenuController header and source files and instead simply using the resulting background image sprite from MainMenuController source file in the GameplayView source file as we needed the same background image used in Main Menu.