Created the Enemy Config file which contains Enemy States, Movement Directions and Enemy Types
Created enum classes to store EnemyStates (PATROLLING, ATTACK, DEAD), EnemyTypes (ZAPPER, SUBZERO, UFO, THUNDER_SNAKE) and Movement Directions (LEFT, RIGHT, DOWN) in the EnemyConfig header file.

Made changes to the EnemyModel header and source files by forward declaring the enum classes created in the EnemyConfig, creating instances of the enum classes and declaring and defining a parameterized constructor with the instance of EnemyType enum class as the parameter and the necessary getter and setter methods for the Enemy States, Types and Movement Directions in the EnemyModel header and source files. Included the EnemyConfig header file in the EnemyModel Source file.

Made changes to the EnemyController header and source files by declaring the destructor, move and initialize functions as virtual in the header file and declaring and defining variables to store the vertical and horizontal movement speed and functions to get random spawn positions of the enemy and to prevent enemies from going out of the screen in the header and source files. Included the EnemyConfig header file in the EnemyController source file.