Implemented movement for the Zapper and Subzero enemies
Declared and defined function to allow SubZero enemy to move downwards in the SubZeroController header and source files.

Declared and defined functions to allow Zapper to move right, left and downwards in the ZapperController header and source files in a similar fashion as the enemy movement which was implemented earlier.

Made changes to the EnemyView header and source files by declaring and defining texture and sprite variables and changing the existing functions for displaying the Zapper and SubZero enemies in the game window. Changed the existing function to initialize the sprites by passing instance of EnemyType enum class as parameter and then within the function creating a switch case to appropriately load the textures, initialize and scale the sprite and display the right enemy based on the current enemy type in the EnemyView source file.