Created the Element Service and the Bunker Controller, Model and View header and source files
The Element Service is responsible for connecting to the MVC files that are responsible for handling and displaying all the elements like bunker in the game.

Created Elements folders for header and source files and created the ElementService header and source file in the folders. Created a namespace called Element and defined the ElementService class in the namespace.

Linked the Element Service to the Service Locator by including the ElementService header file  in the ServiceLocator header file and declaring and defining a pointer to an instance of the ElementService class and using the pointer to call the necessary functions of ElementService in the ServiceLocator header and source files.

Created the Bunker folders inside the Elements folders for header and source files and created the Bunker Model, View and Controller header and source files in the Bunker folders. Created the Bunker namespace in the Element namespace and created the Bunker Model, View and Controller classes.