Declared and defined necessary variables and functions for the Bunker Model, View and Controller classes
Replaced the previously created BunkerModel class with BunkerData struct. Declared and defined variable to store bunker spawning position, default constructor and parameterized constructor with position as parameter to set the bunker position in the BunkerModel header and source files.

Forward declared the BunkerController class and declared and defined pointer to instance of BunkerController class and necessary texture and sprite variables and functions to load the bunker texture, initialize, scale, set the position of the bunker by using the BunkerController pointer to call the function which returns bunker spawning position and display the bunker sprite in the game window in the BunkerView header and source files.

Forward declared the BunkerView class and declared and defined an instance of the BunkerData struct and a pointer to an instance of the BunkerView class and the necessary functions to call the BunkerView functions and to access bunker spawning position from BunkerData in the BunkerController header and source files.

Linked the Bunker View and the Bunker Model to the Bunker Controller by including the BunkerView and BunkerModel header files in the BunkerController source file and including the BunkerController header file in the BunkerView source file.