Created the Sound Service
Created Sounds folder. Created the SoundService header and source files in the Sounds folder. Created the Sound namespace and created the  SoundService class and the enum class for Sound types in the namespace in the SoundService header file. Declared and defined the necessary variables and functions to load and play background music and sound effects in the SoundService header and source files.