Implemented Sound effects and Background Music
Made changes to the SoundService source file by implementing the functions to load and play the sounds and background music in the game.

Linked the Sound Service to the Service Locator by including the SoundService header file in the ServiceLocator header file and creating pointer to instance of SoundService class to call the function to initialize SoundService in the ServiceLocator source file.

Called the functions to play Sound effects and background music after the play button is clicked in the MainMenuController source file.