Created the Bullet Service
The Bullet Service is responsible for displaying and handling the interactions of the bullets fired by the player and enemies in the game.

Created the Bullets folder. Created the BulletService header and source files in the Bullets folder. Created the Bullet namespace and created the BulletService class in the namespace in the header file. Declared and defined the necessary functions in the header and source files.

Created the IProjectile header file. Created the Projectile namespace and created the IProjectile interface in the namespace and declared the pure virtual functions to be implemented by the classes implementing the interface in the header file.