Created the BulletConfig header file and the Bullet Model, View and Controller
Created the BulletConfig header file. Created enum classes for bullet movement directions (UP, DOWN) and for bullet types (LASER_BULLET, TORPEDO, FROST_BULLET).

Created the BulletController, BulletModel and BulletView header files.

The BulletController is responsible for connecting with the BulletModel and BulletView to process bullet movement, interactions and display them in the game window.

Created the BulletController class which implements the IProjectile interface in the BulletController header file. Forward declared the BulletModel , BulletView classes and the BulletType enum class, declared the pointers to instances of the BulletView and BulletModel classes to call functions of their respective classes, declared the functions of the interface to be implemented and necessary additional functions to process and display bullet movement and interactions in the BulletController header file.

Made changes to the BulletService files by keeping Bullet and Projectile as separate namespaces instead of nested and removing the Projectile namespace from the files. Moved the IProjectile interface header file in a new folder called Projectiles.