Created the Bullet Model, View and Controller source files and implemented the BulletModel and BulletView
Created the BulletModel class and forward declared the MovementDirection and BulletType enum classes in the Bullet namespace in the BulletModel header file. Declared and defined the necessary variables and functions to set and get bullet type, bullet speed, bullet direction and bullet position in the BulletModel header header and source files.

Created the BulletView class and forward declared the BulletController class and BulletType enum class in the Bullet namespace in the BulletView header file. Declared and defined the pointer to instance of BulletController class to call functions to get bullet type and position and sprite and texture variables and necessary functions to load the bullet textures and display the bullets in the game window in the BulletView header and source files.