Implemented the Laser Bullet, Frost Bullet and Torpedo Bullet Types
Created a folder Controllers in the Bullets folder. Created the LaserBulletController, TorpedoController and FrostBulletController header and source files in the Controllers folder.

Created a namespace Controller in Bullet namespace and created LaserBulletController class and declared and defined necessary functions to initialize the Laser Bullet in the LaserBulletController header and source files.

Created the TorpedoController class in the Controller namespace and declared and defined variable for torpedo movement speed and necessary functions to initialize torpedo and set torpedo movement speed in the Bullet Model in the TorpedoController header and source files.

Created the FrostBulletController class in the Controller namespace and declared and defined variable for frost bullet movement speed and necessary functions to initialize frost bullet and set frost bullet movement speed in the Bullet Model in the FrostBulletController header and source files.