Implemented the spawning of bullets
Declared and defined vector of IProjectile interface that stores multiple pointers of BulletController class to instances of class of given bullet type in order to call the necessary functions to set the bullet's initial position, movement direction and to update the bullet's status and display it on the game window and necessary functions to create instances of bullets based on current bullet type, to store instances of the bullet types in the vector and to display and destroy bullet type instances.

Linked the Bullet Service to the Service Locator by including the BulletService header file in the ServiceLocator header file and declaring and defining pointer to instance of the BulletService class to call all the necessary functions of the Bullet Service from the ServiceLocator source file.