Implemented firing of bullets by enemies and player
Declared and defined necessary variables and functions to process and implement firing of bullets by enemies with time interval in the EnemyController header and source files. Declared the fire bullet function as virtual to be overridden by the functions of the enemy types in the EnemyController header file.

Declared and defined variable for specifying starting position of the bullet fired by enemy in the BulletModel header file.

Declared and defined the fire bullet function to be overridden in the ZapperController header and source files and called the BulletService function to spawn bullets in the fire bullet function in the ZapperController source file.

Declared and defined variable for specifying subzero enemy type's rate of fire and fire bullet function to be overridden in the SubZeroController header and source files. Implemented the fire bullet function by calling the BulletService spawn bullet function in the SubZeroController source file.

Declared and defined variable to specify starting position of bullet fired by player in PlayerModel header file.

Declared and defined function to fire bullets in the PlayerController header and source files. Implemented the fire bullet function by calling the fire bullet function after checking whether left mouse button has been clicked and then calling the BulletService spawn bullet function in the fire bullet function in the PlayerController source file.