Created the interface for Collectibles
Created a Collectibles folder. Created the ICollectible header file in the folder. Created the Collectible namespace and created the ICollectible interface in the namespace by declaring all the pure virtual functions to initialize, update, render, get position of the collectible and a a function to be called when player picks up the collectible.