Created the Powerup Service, the Powerup Model, View, Controller and Config and created the Outscal Bomb, Triple Laser, Shield and Rapid Fire Powerup Type Controllers
The Powerup Service is responsible for communicating with the Powerup Controller which in turn communicates with the Powerup Model and Powerup View for creating, destroying and displaying the powerups.

Created the Powerups folder. Created the PowerupService, PowerupController, PowerupModel, PowerupView header and source files and the PowerupConfig header file in the folder.

Created the Powerup namespace for all the Powerup related code

Created a PowerType enum class (OUTSCAL_BOMB, SHIELD, RAPID_FIRE, TRIPLE_LASER) in the PowerConfig header file.

Created the PowerupService class in the PowerupService header file. Forward declared the PowerupController class and the PowerupType enum classes. Declared a vector of ICollectible class pointers which stores instances of PowerupType classes depening on current powerup type and the necessary functions to create, spawn and destroy the powerups in the PowerupService header file.

Created the PowerupController class which implements the ICollectible interface in the PowerupController header file. Forward declared PowerupModel and PowerupView classes and the PowerType enum class. Declared and defined pointers to instances of PowerupView and PowerupModel classes and necessary functions to call the appropriate PowerupView and PowerupModel functions to process, update and render powerups in the header and source files.

Created the PowerupModel class in the PowerupModel header file. Declared and defined a variable for powerup movement speed and necessary functions to get and set the powerup position, type and movement speed in the header and source files.

Created the PowerupView class in the PowerupView header file. Forward declared the PowerupController class and PowerupType enum class. Declared and defined pointer to instance of PowerupController class and the texture and sprite variables and necessary functions to load the texture, initialize, scale, update the position of and display the powerup sprite in the game window in the header and source files.

Linked the PowerupController with the PowerupModel and PowerupView by including the PowerupView and PowerupModel header files in the PowerupController source file and including the PowerupController header file in the PowerupView source file.

Created the Controllers folder in the Powerups folder. Created the OutscalBombController, TripleLaserController, ShieldController and RapidFireController header and source files in the Controllers folder. Created the Controller namespace and the respective controller classes which inherit the PowerupController class  in the namespace and declared the PowerupController function to be overridden in the respective header files.