Implemented the Powerup Service and linked it to the Service Locator
Implemented all the functions in the PowerupService source file declared earlier in the PowerupService header file.

Linked the Powerup Service to the Service Locator by including the PowerupService header file and declaring pointer to instance of the PowerupService class in the ServiceLocator header file and defining  the pointer to call appropriate functions of the PowerService class in the ServiceLocator source file.