Created the UFO enemy type
Created the UFOController header and source files. Declared and defined necessary functions in the header and source files to implement movement of UFO (left and right) and to get random Powerup type which is to be spawned when UFO is destroyed by player.

Made necessary changes to the EnemyService and EnemyView source files and EnemyConfig header file to implement spawning and displaying of the UFO in the game window