Created the UI Interface
Created the Interface folder in the UI folder. Created the IUIController header file in the folder. Created the Interface namespace in the UI namespace, created the IUIController interface in the Interface namespace and declared the necessary pure virtual functions in the header file to be implemented by different UI classes implementing this interface.