Made changes to UIService and MainMenuController
Declared and defined the necessary methods to get current UI Controller instance depending on current game state (e.g: MainMenuController if game state is MAIN_MENU) and to assign it to pointer of IUIController interface to call appropriate functions of the UI Controller to update, render and show current UI screen using the pointer in the UIService header and source files.

Defined MainMenuController class as a class which implements the IUIController interface and declared the IUIController interface's methods in the MainMenuController header file to be overridden to update, render and  show the Main Menu Screen.