Created the UI View and Image View
The UI view is responsible for updating state of UI and whether it should be displayed or not and Image View is responsible for displaying the UI and all other components like the Player, enemies, bunkers, bullets, etc of the game in the game window.

Created the UIView header and source files. Created namespace UIElement and created the enum class for UIstates (VISIBLE, HIDDEN) and UIView class in the namespace in the header file. Declared and defined the necessary functions to initialize, update, render, show and hide the game UI in the header and source files.

Created the ImageView header and source files. Created the ImageView class in the UIElement namespace which inherits the UIView class. Declared the necessary texture and sprite variables and the functions to set the properties like texture, scale, position, rotation, opacity of the components in the game and to display them in the game window.

Made changes to the Powerup, Bullet, Bunker, Player and Enemy View header and source files by creating pointer to instance of ImageView class and replacing the original functions to initialize, scale and display components with new functions which uses the pointer to call the necessary functions of ImageView to appropriately load textures, scale, set position of and display the components in the game window.