Created the Text View
Created the TextView header and source files. Created the FontType enum class and the TextView class which inherits the UIView class in the UIElement namespace in the header file. Declared and defined the necessary font variables and the functions to load the font, set the font color, type, text position, text value (content to be displayed) and to display the text in the game window in the header and source files.

Initialized the loading of different font types by calling the appropriate TextView function from the UIService source file.