Created the Button View
The Button View is responsible for creating, displaying and processing interactions on the buttons in the game window.

Created the ButtonView header and source files. Created the ButtonView class which inherits the ImageView class in the UIElement namespace. Declared and defined the necessary callback function pointers (which invoke the functions the pointers are pointing to when button is clicked),  and functions to process button interactions and to call ImageView functions to create, initialize and display the buttons in the header and source files.

Made changes to the MainMenuController header and source files by removing the initial functions and variables to initialize, scale, display and process interactions with the buttons and declaring and defining pointers to instances of the ImageView and ButtonView classes and necessary functions to call the ImageView and ButtonView functions to initialize, scale and display the background image and the buttons and functions to specify and send call back functions of the buttons to be invoked by the ButtonView.